realworld_benchmark/requirements.txt
cudatoolkit==10.0
cudnn==7.6.5
python==3.7.4
python-dateutil==2.8.0
pytorch==1.3
torchvision==0.4.2
pillow==6.1
dgl-cuda10.0==0.4.2
numpy==1.16.4
matplotlib==3.1.0
tensorboard==1.14.0
tensorboardx==1.8S
absl-py
networkx==2.3
scikit-learn==0.21.2
scipy==1.3.0
notebook==6.0.0
h5py==2.9.0
mkl==2019.4
ipykernel==5.1.2
ipython==7.7.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
jupyter==1.0.0
jupyter_client==5.3.1
jupyter_console==6.0.0
jupyter_core==4.5.0
plotly==4.1.1
scikit-image==0.15.0
requests==2.22.0
tqdm==4.43.0
python-hydra
